create the integration test for the post show page
